# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  listingCount?: number;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  active: 'active',
  expired: 'expired',
  deleted: 'deleted',
} as const;

export interface Listing {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  price: number | null;
  currency?: string;
  categoryId: number;
  categoryName: string;
  userId: string;
  userName?: string;
  userPhone: string;
  /** @nullable */
  userAvatarUrl?: string | null;
  imageUrls?: string[];
  /** @nullable */
  location?: string | null;
  status: ListingStatus;
  isFeatured: boolean;
  /** @nullable */
  featuredUntil?: string | null;
  expiresAt: string;
  createdAt: string;
  whatsappUrl?: string;
}

export interface ListingInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @nullable */
  price?: number | null;
  currency?: string;
  categoryId: number;
  userPhone: string;
  imageUrls?: string[];
  location?: string;
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  price?: number | null;
  categoryId?: number;
  userPhone?: string;
  imageUrls?: string[];
  location?: string;
}

export interface ListingsPage {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
  hasMore?: boolean;
}

export interface FeatureInput {
  paymentReference: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  activeListingsCount?: number;
}

export interface UserProfileUpdate {
  name?: string;
  phone?: string;
  avatarUrl?: string;
}

export interface CategoryCount {
  categoryId: number;
  categoryName: string;
  count: number;
}

export interface MarketplaceStats {
  totalListings: number;
  featuredListings: number;
  categoryCounts: CategoryCount[];
}

export type GetListingsParams = {
category?: string;
search?: string;
page?: number;
limit?: number;
featured?: boolean;
};