# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingStatus } from './listingStatus';

export interface Listing {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  price: number | null;
  currency?: string;
  categoryId: number;
  categoryName: string;
  userId: string;
  userName?: string;
  userPhone: string;
  /** @nullable */
  userAvatarUrl?: string | null;
  imageUrls?: string[];
  /** @nullable */
  location?: string | null;
  status: ListingStatus;
  isFeatured: boolean;
  /** @nullable */
  featuredUntil?: string | null;
  expiresAt: string;
  createdAt: string;
  whatsappUrl?: string;
}