# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @nullable */
  price?: number | null;
  currency?: string;
  categoryId: number;
  userPhone: string;
  imageUrls?: string[];
  location?: string;
}