# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  price?: number | null;
  categoryId?: number;
  userPhone?: string;
  imageUrls?: string[];
  location?: string;
}