# lib/api-zod/src/generated/types/listingsPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from './listing';

export interface ListingsPage {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
  hasMore?: boolean;
}