# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface MarketplaceStats {
  totalListings: number;
  featuredListings: number;
  categoryCounts: CategoryCount[];
}