# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  activeListingsCount?: number;
}