# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Marketplace Perú API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  name?: string;
  phone?: string;
  avatarUrl?: string;
}